Fix an increment bug in the mock driver.
While using the mock driver for a while I noticed that my
instances started having weird names.  I eventually figured
out that the mock driver is auto-assigning instance names
based on the instances that already exist.  However, it does
this using the string "succ" method, which means that after
you get to inst9, the next instance is actually insu0.  Fix
it so that we only increment the number part of the name.

Signed-off-by: Chris Lalancette <clalance@redhat.com>