Deal with newly created instances in GoGrid.
From my use of the GoGrid API, it seems there is a
problem searching for newly created instances.  In
particular, trying to do a grid/server/get on them
doesn't actually work until they are more initialized
in the backend.  Deal with this by falling back to a
full listing in the case we get a Bad Request from
the backend.

Signed-off-by: Chris Lalancette <clalance@redhat.com>