Use the instance name as the "ID" for GoGrid.
This is to avoid problems with newly created instances.
GoGrid does not assign a real ID number to an instance
until it has gone through some backend initialization.
That means trying to look up an instance by ID that you
just created will fail, which is pretty non-intuitive
behavior.  Fix this by using the name as the ID, which
probably should be unique enough for our purposes.

Signed-off-by: Chris Lalancette <clalance@redhat.com>