Minor fix for server.rb - affects html case and those drivers where instance 'stop' equivalent to 'destroy'
This is a minor fix in server.rb that affects only those drivers where stopping an instance also means destroying it and where html is the requested output. Basically, in the rackspace case once you stop an instance this gets destroyed *very* quickly. In server.rb and when html is required, all instance actions except destroy cause the driver to request the details of that instance; that is, when you 'reboot' or 'start' you get the details of that specific instance, rather than the index. When you destroy, you dont request details as the instance no longer exists and you automatically get the index.

 However server.rb was not catching the 'stop' case and so when stop actually means destroy, you end up requesting details for an instance you already killed. This problem has only manifested itself in rackspace since they seem to get rid of their instances very quickly (e.g. in other cases like terremark, you might request the instance after you destroy it, and still get it for some time period).